backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
pydantic
python-multipart
aiosqlite
faiss-cpu
pandas
